@@ -397,6 +397,7 @@ public class LibraryContentFragment extends Fragment implements
         recyclerView.addOnScrollListener(scrollListener);
         new FastScrollerBuilder(recyclerView)
                 .setPopupTextProvider(this)
+                .useMd2Style()
                 .build();
 
         // Hide FAB when scrolling up
@@ -1717,7 +1718,7 @@ public class LibraryContentFragment extends Fragment implements
 
     @NonNull
     @Override
-    public CharSequence getPopupText(int position) {
+    public CharSequence getPopupText(@NonNull View view, int position) {
         IAdapter<ContentItem> adapter = getItemAdapter();
         if (null == adapter) return "";
         Content c = adapter.getAdapterItem(position).getContent();
