@@ -38,7 +38,7 @@ public final class DeactivatedPinPreferenceFragment extends Fragment implements 
     @Override
     public void onPinActivateSuccess() {
         Snackbar.make(onSwitch, R.string.app_lock_enable, BaseTransientBottomBar.LENGTH_SHORT).show();
-        HentoidApp.setUnlocked(true); // Now that PIN lock is enabled, the app needs to be marked as currently unlocked to avoid showing an unnecessary PIN dialog at next navigation action
+        HentoidApp.Companion.setUnlocked(true); // Now that PIN lock is enabled, the app needs to be marked as currently unlocked to avoid showing an unnecessary PIN dialog at next navigation action
 
         getParentFragmentManager()
                 .beginTransaction()
