@@ -70,6 +70,7 @@ import me.devsaki.hentoid.util.DebouncerK
 import me.devsaki.hentoid.util.Helper
 import me.devsaki.hentoid.util.Preferences
 import me.devsaki.hentoid.util.Preferences.Constant.*
+import me.devsaki.hentoid.util.Settings
 import me.devsaki.hentoid.util.ThemeHelper
 import me.devsaki.hentoid.util.ToastHelper
 import me.devsaki.hentoid.util.exception.ContentNotProcessedException
@@ -101,9 +102,10 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
     private lateinit var adapter: ImagePagerAdapter
     private lateinit var llm: PrefetchLinearLayoutManager
     private lateinit var pageSnapWidget: PageSnapWidget
-    private val listener = OnSharedPreferenceChangeListener { _: SharedPreferences, key: String ->
-        onSharedPreferenceChanged(key)
-    }
+    private val prefsListener =
+        OnSharedPreferenceChangeListener { _: SharedPreferences, key: String ->
+            onSharedPreferenceChanged(key)
+        }
     private var viewModel: ReaderViewModel? = null
     private var absImageIndex = -1 // Absolute (book scale) 0-based image index
 
@@ -139,6 +141,8 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
     private lateinit var indexRefreshDebouncer: DebouncerK<Int>
     private lateinit var processPositionDebouncer: DebouncerK<Pair<Int, Int>>
     private lateinit var rescaleDebouncer: DebouncerK<Float>
+    private lateinit var adapterRescaleDebouncer: DebouncerK<Float>
+    private var firstZoom = true
 
     // Starting index management
     private var isComputingImageList = false
@@ -182,8 +186,19 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
         }
         rescaleDebouncer = DebouncerK(this.lifecycleScope, 100) { scale: Float ->
             adapter.multiplyScale(scale)
+            if (!firstZoom)
+                context?.let {
+                    ToastHelper.toast(it, R.string.percent_no_digits, scale * 100.0)
+                }
+            else firstZoom = false
         }
-        Preferences.registerPrefsChangedListener(listener)
+        adapterRescaleDebouncer = DebouncerK(this.lifecycleScope, 100) { scale: Float ->
+            context?.let {
+                ToastHelper.toast(it, R.string.percent_no_digits, scale * 100.0)
+            }
+        }
+        Preferences.registerPrefsChangedListener(prefsListener)
+        Settings.registerPrefsChangedListener(prefsListener)
     }
 
     @SuppressLint("NonConstantResourceId")
@@ -249,6 +264,7 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
         slideshowSliderDebouncer.clear()
         processPositionDebouncer.clear()
         rescaleDebouncer.clear()
+        adapterRescaleDebouncer.clear()
         navigator.clear()
         binding?.recyclerView?.adapter = null
         binding = null
@@ -338,7 +354,8 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
     }
 
     override fun onDestroy() {
-        Preferences.unregisterPrefsChangedListener(listener)
+        Preferences.unregisterPrefsChangedListener(prefsListener)
+        Settings.unregisterPrefsChangedListener(prefsListener)
         if (this::adapter.isInitialized) {
             adapter.setRecyclerView(null)
             adapter.destroy()
@@ -408,6 +425,7 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
                 )
             }
             recyclerView.requestFocus()
+            // Scale listener from the ImageView, incl. top to bottom browsing
             recyclerView.setOnScaleListener { scale ->
                 if (LinearLayoutManager.HORIZONTAL == llm.orientation) {
                     pageSnapWidget.let {
@@ -426,6 +444,10 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
                     return false
                 }
             })
+            // Scale listener from the SSIV
+            adapter.setOnScaleListener { position, scale ->
+                if (position == absImageIndex) adapterRescaleDebouncer.submit(scale)
+            }
 
             val tapZoneScale = if (Preferences.isReaderTapToTurn2x()) 2 else 1
             val onHorizontalZoneTapListener = OnZoneTapListener(
@@ -898,6 +920,7 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
             }
 
             navigator.updatePageControls()
+            firstZoom = true
         }
     }
 
@@ -920,7 +943,7 @@ class ReaderPagerFragment : Fragment(R.layout.fragment_reader_pager),
             Preferences.Key.VIEWER_ZOOM_TRANSITIONS, Preferences.Key.VIEWER_SEPARATING_BARS, Preferences.Key.VIEWER_AUTO_ROTATE
             -> onUpdateImageDisplay(true)
 
-            Preferences.Key.VIEWER_IMAGE_DISPLAY, Preferences.Key.VIEWER_RENDERING
+            Preferences.Key.VIEWER_IMAGE_DISPLAY, Preferences.Key.VIEWER_RENDERING, Settings.Key.READER_COLOR_DEPTH
             -> onUpdateImageDisplay(false)
 
             Preferences.Key.VIEWER_SWIPE_TO_FLING -> onUpdateSwipeToFling()
