@@ -3,30 +3,43 @@ package me.devsaki.hentoid.workers
 import android.content.Context
 import android.graphics.BitmapFactory
 import android.graphics.Point
+import androidx.core.net.toUri
 import androidx.documentfile.provider.DocumentFile
 import androidx.work.Data
 import androidx.work.WorkerParameters
 import com.bumptech.glide.Glide
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import me.devsaki.hentoid.R
+import me.devsaki.hentoid.ai_upscale.AiUpscaler
 import me.devsaki.hentoid.database.CollectionDAO
 import me.devsaki.hentoid.database.ObjectBoxDAO
 import me.devsaki.hentoid.database.domains.Content
 import me.devsaki.hentoid.database.domains.ImageFile
 import me.devsaki.hentoid.notification.transform.TransformCompleteNotification
 import me.devsaki.hentoid.notification.transform.TransformProgressNotification
+import me.devsaki.hentoid.util.Helper
 import me.devsaki.hentoid.util.file.FileHelper
 import me.devsaki.hentoid.util.image.ImageHelper
 import me.devsaki.hentoid.util.image.ImageTransform
 import me.devsaki.hentoid.util.notification.Notification
+import timber.log.Timber
+import java.io.File
+import java.nio.ByteBuffer
 import java.time.Instant
 import java.util.concurrent.atomic.AtomicInteger
+
 
 class TransformWorker(context: Context, parameters: WorkerParameters) :
     BaseWorker(context, parameters, R.id.transform_service, null) {
 
     private val dao: CollectionDAO
+    private var upscaler: AiUpscaler? = null
+
     private var totalItems = 0
     private var nbOK = 0
     private var nbKO = 0
@@ -46,6 +59,7 @@ class TransformWorker(context: Context, parameters: WorkerParameters) :
 
     override fun onClear() {
         dao.cleanup()
+        upscaler?.cleanup()
 
         // Reset Glide cache as it gets confused by the resizing
         Glide.get(applicationContext).clearDiskCache()
@@ -58,12 +72,19 @@ class TransformWorker(context: Context, parameters: WorkerParameters) :
         require(paramsStr != null)
         require(paramsStr.isNotEmpty())
 
-        val moshi = Moshi.Builder()
-            .addLast(KotlinJsonAdapterFactory())
-            .build()
+        val moshi = Moshi.Builder().addLast(KotlinJsonAdapterFactory()).build()
 
         val params = moshi.adapter(ImageTransform.Params::class.java).fromJson(paramsStr)
         require(params != null)
+
+        if (3 == params.resizeMethod) { // AI upscale
+            upscaler = AiUpscaler()
+            upscaler!!.init(
+                applicationContext.resources.assets,
+                "realsr/models-nose/up2x-no-denoise.param",
+                "realsr/models-nose/up2x-no-denoise.bin"
+            )
+        }
 
         transform(contentIds, params)
     }
@@ -91,15 +112,13 @@ class TransformWorker(context: Context, parameters: WorkerParameters) :
             FileHelper.getDocumentFromTreeUriString(applicationContext, content.storageUri)
         val images = content.imageList
         if (contentFolder != null) {
-            val imagesWithoutChapters = images
-                .filter { i -> null == i.linkedChapter }
-                .filter { i -> i.isReadable }
+            val imagesWithoutChapters =
+                images.filter { i -> null == i.linkedChapter }.filter { i -> i.isReadable }
             transformChapter(imagesWithoutChapters, contentFolder, params)
 
-            val chapteredImgs = images
-                .filterNot { i -> null == i.linkedChapter }
-                .filter { i -> i.isReadable }
-                .groupBy { i -> i.linkedChapter!!.id }
+            val chapteredImgs =
+                images.filterNot { i -> null == i.linkedChapter }.filter { i -> i.isReadable }
+                    .groupBy { i -> i.linkedChapter!!.id }
 
             chapteredImgs.forEach {
                 transformChapter(it.value, contentFolder, params)
@@ -117,12 +136,11 @@ class TransformWorker(context: Context, parameters: WorkerParameters) :
     }
 
     private fun transformChapter(
-        imgs: List<ImageFile>,
-        contentFolder: DocumentFile,
-        params: ImageTransform.Params
+        imgs: List<ImageFile>, contentFolder: DocumentFile, params: ImageTransform.Params
     ) {
         val nbManhwa = AtomicInteger(0)
         params.forceManhwa = false
+
         imgs.forEach {
             transformImage(it, contentFolder, params, nbManhwa, imgs.size)
             if (isStopped) return
@@ -145,22 +163,68 @@ class TransformWorker(context: Context, parameters: WorkerParameters) :
         val rawData = FileHelper.getInputStream(applicationContext, sourceFile).use {
             return@use it.readBytes()
         }
-        val isLossless = ImageHelper.isImageLossless(rawData)
-        val sourceName = sourceFile.name ?: ""
-        val options = BitmapFactory.Options()
-        options.inJustDecodeBounds = true
-        BitmapFactory.decodeByteArray(rawData, 0, rawData.size, options)
-        val isManhwa = options.outHeight * 1.0 / options.outWidth > 3
+        val metadataOpts = BitmapFactory.Options()
+        metadataOpts.inJustDecodeBounds = true
 
-        if (isManhwa) nbManhwa.incrementAndGet()
-        params.forceManhwa = nbManhwa.get() * 1.0 / nbPages > 0.9
+        val targetData: ByteArray
+        if (upscaler != null) {
+            val cacheDir =
+                FileHelper.getOrCreateCacheFolder(applicationContext, "upscale") ?: return
+            val outputFile = File(cacheDir, "upscale.png")
+            val progress = ByteBuffer.allocateDirect(1)
+            val dataIn = ByteBuffer.allocateDirect(rawData.size)
+            dataIn.put(rawData)
 
-        val targetData = ImageTransform.transform(rawData, params)
+            upscaler?.let {
+                try {
+                    CoroutineScope(Dispatchers.Default).launch {
+                        val res = withContext(Dispatchers.Default) {
+                            it.upscale(
+                                dataIn, outputFile.absolutePath, progress
+                            )
+                        }
+                        // Fail => exit immediately
+                        if (res != 0) progress.put(0, 100)
+                    }
+
+                    // Poll while processing
+                    val intervalSeconds = 3
+                    var iterations = 0
+                    while (iterations < 180 / intervalSeconds) { // max 3 minutes
+                        Helper.pause(intervalSeconds * 1000)
+                        val p = progress.get(0)
+                        // TODO use that to update notification progress to a finer degree
+                        Timber.i("progress : %s%%", p)
+                        if (p >= 100) break
+                        iterations++
+                    }
+                } finally {
+                    // can't recycle ByteBuffer dataIn
+                }
+            }
+
+            FileHelper.getInputStream(applicationContext, outputFile.toUri()).use { input ->
+                targetData = input.readBytes()
+            }
+        } else { // regular resize
+            BitmapFactory.decodeByteArray(rawData, 0, rawData.size, metadataOpts)
+            val isManhwa = metadataOpts.outHeight * 1.0 / metadataOpts.outWidth > 3
+
+            if (isManhwa) nbManhwa.incrementAndGet()
+            params.forceManhwa = nbManhwa.get() * 1.0 / nbPages > 0.9
+
+            targetData = ImageTransform.transform(rawData, params)
+        }
         if (targetData == rawData) return // Unchanged picture
 
-        BitmapFactory.decodeByteArray(targetData, 0, targetData.size, options)
-        val targetDims = Point(options.outWidth, options.outHeight)
-        val targetMime = ImageTransform.determineEncoder(isLossless, targetDims, params).mimeType
+        // Save transformed image data back to original image file
+        val isLossless = ImageHelper.isImageLossless(rawData)
+        val sourceName = sourceFile.name ?: ""
+
+        BitmapFactory.decodeByteArray(targetData, 0, targetData.size, metadataOpts)
+        val targetDims = Point(metadataOpts.outWidth, metadataOpts.outHeight)
+        val targetMime =
+            ImageTransform.determineEncoder(isLossless, targetDims, params).mimeType
         val targetName = img.name + "." + FileHelper.getExtensionFromMimeType(targetMime)
         val newFile = sourceName != targetName
 
